requirements.txt
langchain-openai
langchain_community
langchain_core
langchain
streamlit
python-dotenv
langchain-groq
pymupdf
faiss-cpu
langchain_chroma
duckdb
langchain_huggingface
chromadb
pypdf
langchain-text-splitters
openai
pandas
validators==0.34.0
youtube_transcript_api
unstructured
wikipedia
